refactor: move ThemePreview to components directory
